change animation leaderboard cast window
